third_party/blink/web_tests/platform/generic/fast/sub-pixel/replaced-element-baseline-expected.txt

PASS document.getElementById('a').getBoundingClientRect().top is 0
PASS document.getElementById('b').getBoundingClientRect().top is 0
PASS successfullyParsed is true

TEST COMPLETE
